+++ Templates/Partials/MethodParameter.swift
@@ -16,7 +16,10 @@ func stringify(parameters: [MethodParameter], printing: MethodParameterPrinting)
         case .call:
             mapping = { $0.name }
         case .definition:
-            mapping = { "\($0.name): \($0.typeName)" }
+            mapping = {
+                "\($0.name): \($0.typeName)"
+                + ($0.defaultValue.map { " = " + $0 } ?? "")
+            }
     }
 
     return parameters